pyproject.toml
[tool.poetry]
name = "db-data-in-out"
version = "0.1.0"
description = ""
authors = ["MDiRocco <marc.dirocco@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
typer = "^0.12.5"
sqlalchemy = "^2.0.36"
psycopg2 = "^2.9.10"
openpyxl = "^3.1.5"
pyyaml = "^6.0.2"
pandas = "^2.2.3"
fire = "^0.7.0"
wemake-python-styleguide = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
